backend/requirements.txt
# Core Framework & Server
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.9

# HTTP Client
httpx==0.27.2

# Telegram & Crypto
pyrogram==2.0.106
tgcrypto==1.2.5
python-telegram-bot==21.1.1

# WebSocket Real-Time Features
python-socketio[asyncio]==5.11.0

# Distributed Caching & Rate Limiting
redis==5.0.3

# Configuration Management
pydantic-settings==2.2.1

# JWT Handling
python-jose[cryptography]==3.3.0

# Password Hashing
passlib[bcrypt]==1.7.4

# Image Processing (Thumbnails)
Pillow==10.3.0

# Async File Operations
aiofiles==23.2.1

# Memory Monitoring
psutil==5.9.8
